+++ crates/orbital/src/osc/primary.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 pub struct PrimaryOsc {
     ///Base frequency multiplier. This basically means if a note @ 440Hz is played, and this is 0.5, then
     /// the primary oscillator has a base frequency of 220Hz
-    pub speed_index: i32,
+    pub speed_index: f32,
     pub volume: f32,
     pub is_on: bool,
 }
@@ -14,14 +14,14 @@ pub struct PrimaryOsc {
 impl PrimaryOsc {
     #[inline(always)]
     pub fn freq(&self, base_frequency: f32) -> f32 {
-        base_frequency * 2.0f32.powf(self.speed_index as f32)
+        base_frequency * 2.0f32.powf(self.speed_index)
     }
 }
 
 impl Default for PrimaryOsc {
     fn default() -> Self {
         PrimaryOsc {
-            speed_index: 0,
+            speed_index: 0.0,
             volume: 0.0,
             is_on: false,
         }
